IA SDK updated to 2.4
