+++ README.md
@@ -1,4 +1,4 @@
-# IndoorAtlas Cordova Plugin 1.1 with IndoorAtlas SDK 2.3
+# IndoorAtlas Cordova Plugin 1.2 with IndoorAtlas SDK 2.4
 
 This Cordova plugin provides information about the device's location inside finger printed buildings and uses the [IndoorAtlas location APIs](http://docs.indooratlas.com/cordova/api-documentation).
