--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Headers
@@ -1,1 +1,0 @@
-Versions/Current/Headers
