--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/IndoorAtlas
@@ -1,1 +1,0 @@
-Versions/Current/IndoorAtlas
