--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Modules
@@ -1,1 +1,0 @@
-Versions/Current/Modules
