--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Resources
@@ -1,1 +1,0 @@
-Versions/Current/Resources
