--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Versions/2.3.1-57/Headers/IndoorAtlas.h
@@ -1,7 +1,0 @@
-// IndoorAtlas iOS SDK
-// IndoorAtlas.h
-
-#import <IndoorAtlas/IAFloor.h>
-#import <IndoorAtlas/IAFloorPlan.h>
-#import <IndoorAtlas/IALocationManager.h>
-#import <IndoorAtlas/IAResourceManager.h>
