--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Versions/2.3.1-57/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module IndoorAtlas { umbrella header "IndoorAtlas.h"
-header "IALocationManager.h"
-header "IAResourceManager.h"
-header "IAFloorPlan.h"
-header "IAFloor.h"
-export * module * { export * } }
