--- src/ios/IndoorAtlas2.3/IndoorAtlas.framework/Versions/Current
@@ -1,1 +1,0 @@
-2.3.1-57
