+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
