+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/IndoorAtlas
@@ -1,0 +1,1 @@
+Versions/Current/IndoorAtlas
