+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/Modules
@@ -1,0 +1,1 @@
+Versions/Current/Modules
