+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/Resources
@@ -1,0 +1,1 @@
+Versions/Current/Resources
