+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/Versions/2.4.1-66/Headers/IndoorAtlas.h
@@ -1,0 +1,7 @@
+// IndoorAtlas iOS SDK
+// IndoorAtlas.h
+
+#import <IndoorAtlas/IAFloor.h>
+#import <IndoorAtlas/IAFloorPlan.h>
+#import <IndoorAtlas/IALocationManager.h>
+#import <IndoorAtlas/IAResourceManager.h>
