+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/Versions/2.4.1-66/Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module IndoorAtlas { umbrella header "IndoorAtlas.h"
+header "IALocationManager.h"
+header "IAResourceManager.h"
+header "IAFloorPlan.h"
+header "IAFloor.h"
+export * module * { export * } }
