+++ src/ios/IndoorAtlas2.4/IndoorAtlas.framework/Versions/Current
@@ -1,0 +1,1 @@
+2.4.1-66
